Move KafkaTopicsExistsTest
